Implemented Spell Behavior
Updated the wizard class to fire off loaded spell 1 on the wand. Spells now all fire using a direction supplied to them, the functionality of which will need to be baked into the Command structure.
Will need to revise the structure of Spell, as many of the fields are unnecessary.